ofxsPixelProcessor.h: fix bad cast (multiplication before casting)
fixes https://github.com/NatronGitHub/Natron/issues/456

+++ ofxsPixelProcessor.h
@@ -134,7 +134,7 @@ getPixelAddress(void* pixelData,
     if ( ( x < bounds.x1) || ( x >= bounds.x2) || ( y < bounds.y1) || ( y >= bounds.y2) || ( pixelBytes == 0) ) {
         return 0;
     }
-    char *pix = (char *) ( ( (char *) pixelData ) + (size_t)((y - bounds.y1) * rowBytes) );
+    char *pix = (char *) ( ( (char *) pixelData ) + (size_t)(y - bounds.y1) * rowBytes );
     pix += (x - bounds.x1) * pixelBytes;
 
     return (void *) pix;
@@ -158,7 +158,7 @@ getPixelAddress(const void* pixelData,
              ( x < bounds.x1) || ( x >= bounds.x2) || ( y < bounds.y1) || ( y >= bounds.y2) || ( pixelBytes == 0) ) ) {
         return 0;
     }
-    char *pix = (char *) ( ( (char *) pixelData ) + (size_t)(y - bounds.y1) * rowBytes );
+    char *pix = (char *) ( ( (char *) pixelData ) + (size_t)(y - bounds.y1) * (size_t)rowBytes );
     pix += (x - bounds.x1) * pixelBytes;
 
     return (void *) pix;
@@ -300,7 +300,7 @@ protected:
             return 0;
         }
 
-        char *pix = (char *) ( ( (char *) _dstPixelData ) + (size_t)((y - _dstBounds.y1) * _dstRowBytes) );
+        char *pix = (char *) ( ( (char *) _dstPixelData ) + (size_t)(y - _dstBounds.y1) * _dstRowBytes );
         pix += (x - _dstBounds.x1) * _dstPixelBytes;
 
         return (void *) pix;
@@ -446,7 +446,7 @@ protected:
             }
         }
 
-        char *pix = (char *) ( ( (char *) _srcPixelData ) + (size_t)((y - _srcBounds.y1) * _srcRowBytes) );
+        char *pix = (char *) ( ( (char *) _srcPixelData ) + (size_t)(y - _srcBounds.y1) * _srcRowBytes );
         pix += (x - _srcBounds.x1) * _srcPixelBytes;
 
         return (void *) pix;
